+++ src/utils/client/switchLang.test.ts
@@ -2,7 +2,7 @@ import { setCookie } from '@hermeslabs/utils';
 import { changeLanguage } from 'i18next';
 import { describe, expect, it, vi } from 'vitest';
 
-import { LOBE_LOCALE_COOKIE } from '@/const/locale';
+import { HERMES_LOCALE_COOKIE, LEGACY_LOBE_LOCALE_COOKIE } from '@/const/locale';
 import { LocaleMode } from '@/types/locale';
 
 import { switchLang } from './switchLang';
@@ -26,7 +26,8 @@ describe('switchLang', () => {
 
     expect(changeLanguage).toHaveBeenCalledWith(locale);
     expect(document.documentElement.lang).toBe(locale);
-    expect(setCookie).toHaveBeenCalledWith(LOBE_LOCALE_COOKIE, locale, 365);
+    expect(setCookie).toHaveBeenNthCalledWith(1, HERMES_LOCALE_COOKIE, locale, 365);
+    expect(setCookie).toHaveBeenNthCalledWith(2, LEGACY_LOBE_LOCALE_COOKIE, locale, 365);
   });
 
   it('should change language based on navigator.language when locale is "auto"', () => {
@@ -37,6 +38,7 @@ describe('switchLang', () => {
 
     expect(changeLanguage).toHaveBeenCalledWith(navigatorLanguage);
     expect(document.documentElement.lang).toBe(navigatorLanguage);
-    expect(setCookie).toHaveBeenCalledWith(LOBE_LOCALE_COOKIE, undefined, 365);
+    expect(setCookie).toHaveBeenNthCalledWith(1, HERMES_LOCALE_COOKIE, undefined, 365);
+    expect(setCookie).toHaveBeenNthCalledWith(2, LEGACY_LOBE_LOCALE_COOKIE, undefined, 365);
   });
 });